refactor(client): extract Chrono + File Meta tabs, delete loadCurrentTab (A17.9b-6g)
The last two record tabs, in one PR (the pattern is proven three times over).

- chrono-tab.component: follows the memories/edges pattern (semantic pill). Pinned
  quirks preserved — create resolves a `__custom__` kind while inline-edit sends the
  kind verbatim, and neither create nor delete refreshes stats (so it has no `mutated`
  output).
- filemeta-tab.component: the odd tab — no create form (records come from ingested
  files), the files API for save/delete (delete by path, which DOES refresh stats),
  the shared `fm` memory/chrono pickers on EntityRefPicker, retryFileEmbedding, and no
  semantic mode (client-side filter). Navigating to the Files tab is shell nav, so it
  is an `openInManager` output the shell handles.

With all five record tabs now self-loading components, the shell's loadCurrentTab
dispatcher is deleted, and with it the dead per-tab skip/filter state and the
onFilterChange/applyFilter/clearFilter/retryCurrentTab helpers; setTab/selectSpace
shrink to nav + the store's cross-tab search resets. The shell now owns only
navigation: space chips, activeTab/activeSpaceId, the tab bar with count badges, and
it renders <app-*-tab> + <app-record-drawer/>. Its now-unused component imports
(PropertiesView/EntitySearch/TagInput/ErrorState/RecordFilterBar) and APIs
(BrainApi/FilesApi/ToastService) were dropped.

The 6b chrono/filemeta characterization cases moved into the two component specs; the
now-empty brain.component.records.spec.ts was deleted.

brain.component.ts 1402 -> 637 (3701 -> 637 across A17.9b). Client suite 200 -> 208.
tsc --noUnusedLocals clean; prod AOT build warning-free bar the pre-existing qrcode
bailout. Only the dead record-table CSS remains in the shell styles — swept up next.
